requirements.txt
# Render-compatible requirements.txt - No errors
Flask==3.0.3
Flask-CORS==4.0.1
gunicorn==22.0.0
pymongo==4.8.0
bcrypt==4.1.3
PyJWT==2.9.0
python-dateutil==2.9.0
python-dotenv==1.0.1
requests==2.32.3
openai
pillow==10.4.0
PyPDF2==3.0.1
python-docx==1.1.2
numpy==1.24.3
opencv-python-headless==4.8.1.78
matplotlib==3.8.4
aiohttp==3.9.5
protobuf>=3.19.5,<5.0.0

# PaddleOCR dependencies for medication processing
paddlepaddle==2.5.2
paddleocr==2.7.0.3

# Note: Heavy dependencies removed to fit 512MB limit
# - sentence-transformers (saves ~300MB)
# - PyMuPDF (Python 3.13 incompatible)
# - scipy (saves ~50MB)
# - voice services (saves ~200MB)